+++ core/registration.go
@@ -36,7 +36,7 @@ func Register(r Registration) (husk.Recorder, error) {
 
 	//TODO: Make dynamic
 	//Should provide only basic Resources, the rest will be unlocked later
-	user := prime.NewUser(r.Name, r.Email, []byte(r.Password), false, contc, nil)
+	user := prime.NewUser(r.Name, r.Email, r.Password, false, contc, nil)
 
 	rec := ctx.Users.Create(user.(prime.User))
 	defer ctx.Users.Save()