+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 ## Mango-Secure
 
+# 1.0.4
+
+* Fixed Resource path.
+
+# 1.0.3
+
+* Fixed Update Resource object param
+
+# 1.0.2
+
+* Add Resource functions and bodies
+
 # 1.0.1
 
 * Add Profile functions and bodies